# ISSUE-3618: Improved performance of arithmetic plus functions