--- a/common/functions/src/scalars/arithmetics/arithmetic_intdiv.rs
+++ b/common/functions/src/scalars/arithmetics/arithmetic_intdiv.rs
@@ -29,8 +29,12 @@ impl ArithmeticIntDivFunction {
     }
 
     pub fn desc() -> FunctionDescription {
-        FunctionDescription::creator(Box::new(Self::try_create_func))
-            .features(FunctionFeatures::default().deterministic().monotonicity())
+        FunctionDescription::creator(Box::new(Self::try_create_func)).features(
+            FunctionFeatures::default()
+                .deterministic()
+                .monotonicity()
+                .num_arguments(2),
+        )
     }
 
     pub fn get_monotonicity(_args: &[Monotonicity]) -> Result<Monotonicity> {