--- a/common/functions/src/scalars/arithmetics/binary_arithmetic.rs
+++ b/common/functions/src/scalars/arithmetics/binary_arithmetic.rs
@@ -21,11 +21,6 @@ use common_datavalues::DataValueArithmeticOperator;
 use common_exception::Result;
 
 use super::arithmetic::ArithmeticTrait;
-/*use crate::scalars::ArithmeticDivFunction;
-use crate::scalars::ArithmeticIntDivFunction;
-use crate::scalars::ArithmeticMinusFunction;
-use crate::scalars::ArithmeticModuloFunction;
-use crate::scalars::ArithmeticMulFunction;*/
 use crate::scalars::ArithmeticPlusFunction;
 use crate::scalars::Function;
 use crate::scalars::Monotonicity;
@@ -80,11 +75,6 @@ where T: ArithmeticTrait + Clone + Sync + Send + 'static
         match self.op {
             Plus => ArithmeticPlusFunction::get_monotonicity(args),
             _ => unimplemented!(),
-            /* Minus => ArithmeticMinusFunction::get_monotonicity(args),
-            Mul => ArithmeticMulFunction::get_monotonicity(args),
-            Div => ArithmeticDivFunction::get_monotonicity(args),
-            IntDiv => ArithmeticIntDivFunction::get_monotonicity(args),
-            Modulo => ArithmeticModuloFunction::get_monotonicity(args),*/
         }
     }
 }