… Cargo.lock
… common/functions/Cargo.toml
… common/functions/src/scalars/arithmetics/interval.rs
… common/functions/src/scalars/arithmetics/macros.rs
… common/functions/src/scalars/arithmetics/mod.rs
… common/functions/src/scalars/arithmetics/utils.rs
… common/functions/src/scalars/comparisons/comparison.rs
… common/functions/src/scalars/comparisons/comparison_eq.rs
… common/functions/src/scalars/comparisons/comparison_gt.rs
… common/functions/src/scalars/comparisons/comparison_gt_eq.rs
… common/functions/src/scalars/comparisons/comparison_like.rs
… common/functions/src/scalars/comparisons/comparison_lt.rs
… common/functions/src/scalars/comparisons/comparison_lt_eq.rs
… common/functions/src/scalars/comparisons/comparison_not_eq.rs
… common/functions/src/scalars/comparisons/comparison_not_like.rs
… common/functions/src/scalars/conditionals/if.rs
… common/functions/src/scalars/dates/date.rs
… common/functions/src/scalars/dates/interval_function.rs
… common/functions/src/scalars/dates/number_function.rs
… common/functions/src/scalars/dates/round_function.rs
… common/functions/src/scalars/dates/week_date.rs
… common/functions/src/scalars/expressions/cast.rs
… common/functions/src/scalars/expressions/expression.rs
… common/functions/src/scalars/function.rs
… common/functions/src/scalars/function_alias.rs
… common/functions/src/scalars/function_column.rs
… common/functions/src/scalars/function_factory.rs
… common/functions/src/scalars/function_literal.rs
… common/functions/src/scalars/hashes/sha2hash.rs
… common/functions/src/scalars/hashes/siphash.rs