I hereby agree to the terms of the CLA available at: https://databend.rs/policies/cla/

## Summary

Move arithmetic from datavalues to functions.
The type is no longer converted before the calculation, but is converted during the calculation.
Move num_argments and variadic_arguments from Function trait to FunctionFeatures.

## Changelog

- Improvement
- Performance Improvement

## Related Issues

Fixes #3618 

## Test Plan

Unit Tests

Stateless Tests