@zhyass 2021-12-23
Performance on Apple M1 16G
Before:
```
mysql> SELECT count() FROM numbers(10000000000) WHERE NOT ignore(number + number);
+-------------+
| count()     |
+-------------+
| 10000000000 |
+-------------+
1 row in set (1.35 sec)
Read 10000000000 rows, 80 GB in 1.345 sec., 7.43 billion rows/sec., 59.47 GB/sec.

mysql> SELECT count() FROM numbers(10000000000) WHERE NOT ignore(toUInt32(number) + number);
+-------------+
| count()     |
+-------------+
| 10000000000 |
+-------------+
1 row in set (2.08 sec)
Read 10000000000 rows, 80 GB in 2.077 sec., 4.81 billion rows/sec., 38.51 GB/sec.

mysql> SELECT count() FROM numbers(10000000000) WHERE NOT ignore(toUInt32(number) + toUInt16(number));
+-------------+
| count()     |
+-------------+
| 10000000000 |
+-------------+
1 row in set (3.22 sec)
Read 10000000000 rows, 80 GB in 3.227 sec., 3.1 billion rows/sec., 24.79 GB/sec.
```

After:
```
mysql> SELECT count() FROM numbers(10000000000) WHERE NOT ignore(number + number);
+-------------+
| count()     |
+-------------+
| 10000000000 |
+-------------+
1 row in set (1.35 sec)
Read 10000000000 rows, 80 GB in 1.350 sec., 7.41 billion rows/sec., 59.26 GB/sec.

mysql> SELECT count() FROM numbers(10000000000) WHERE NOT ignore(toUInt32(number) + number);
+-------------+
| count()     |
+-------------+
| 10000000000 |
+-------------+
1 row in set (1.65 sec)
Read 10000000000 rows, 80 GB in 1.645 sec., 6.08 billion rows/sec., 48.64 GB/sec.

mysql> SELECT count() FROM numbers(10000000000) WHERE NOT ignore(toUInt32(number) + toUInt16(number));
+-------------+
| count()     |
+-------------+
| 10000000000 |
+-------------+
1 row in set (2.35 sec)
Read 10000000000 rows, 80 GB in 2.345 sec., 4.26 billion rows/sec., 34.12 GB/sec.
```